+++ bin/downsample_logging.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-# Copyright 2019 Jetperch LLC
+# Copyright 2019-2024 Jetperch LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -18,20 +18,23 @@
 r"""
 Capture downsampled data to a comma-separate values (CSV) file.
 
-This script collects full-rate 2 MHz data from Joulescope, downsamples the data
-to 2 Hz, and then records the downsampled data to a CSV file.  The CSV files 
-are automatically named and stored under the Documents/joulescope directory. 
-On Windows, this will typically be:
+This script records statistics data to CSV files, one per Joulescope.
+The CSV files are automatically named and stored under the
+Documents/joulescope directory.  On Windows, this will typically be:
 
-    C:\Users\{user_name}\Documents\joulescope\jslog_{YYYYMMDD_hhmmss}_{pid}_{serial_number}.csv
+    C:\Users\{user_name}\Documents\joulescope
+
+Each file is name like:
+
+    jslog_{YYYYMMDD_hhmmss}_{pid}_{model}_{serial_number}.csv
 
 The ".csv" file contains the capture data with columns:
 
-    time,current,voltage,power,charge,energy
+    time,current,voltage,power,charge,energy,current_min,current_max
 
 All values are in the International System of Units (SI):
 
-    seconds,amperes,volts,watts,coulombs,joules
+    seconds,amperes,volts,watts,coulombs,joules,amperes,amperes
 
 The script also creates a ".txt" file which contains the state information
 for the logging session.
@@ -49,7 +52,8 @@ This implementation handles failure modes including:
 * Temporary loss of system power (using --resume option)
 * Host computer reset (using --resume option)
 
-For long-term logging, even 2 Hz downsampled data can still create too much data:
+For very long-term logging, even 2 Hz downsampled data may still create
+too much data:
 
     2 lines/second * (60 seconds/minute * 60 minutes/hour * 24 hours/day) = 
     172800 lines / day
@@ -73,16 +77,9 @@ a 3.3V supply and 1000 Ω resistive load (10.9 mW):
     180.0513701,0.00329558,3.2998,0.0108748,0.593167,1.95733
     240.0502210,0.00329565,3.2998,0.0108751,0.790906,2.60984
     300.0581367,0.00329583,3.2997,0.0108751,0.988656,3.26234
-
-
-List of requested future features:
-* Specify save path (but can't break resume function!)
-* Add higher sample rate (see forum https://forum.joulescope.com/t/joulescope-data-logging/130/10)
-
 """
 
-import joulescope
-from joulescope.data_recorder import DataRecorder
+from pyjoulescope_driver import Driver, time64
 import signal
 import argparse
 import time
@@ -90,35 +87,23 @@ import sys
 import os
 import datetime
 import logging
-import json
-import weakref
 import numpy as np
+import queue
+import threading
 
+
+MAX_SAMPLES = 1000000000 / 5  # limit to 1 GB of RAM
+CSV_SEEK = 4096
+LAST_INITIALIZE = (None, 0.0, 0.0, 0.0, 0.0, 0.0)
+USER_NOTIFY_INTERVAL_S = 10.0
+FLOAT_MAX = np.finfo(float).max
 
 try:
     from win32com.shell import shell, shellcon
     DOCUMENTS_PATH = shell.SHGetFolderPath(0, shellcon.CSIDL_PERSONAL, None, 0)
     BASE_PATH = os.path.join(DOCUMENTS_PATH, 'joulescope')
-
 except Exception:
     BASE_PATH = os.path.expanduser('~/Documents/joulescope')
-
-
-MAX_SAMPLES = 1000000000 / 5  # limit to 1 GB of RAM
-CSV_SEEK = 4096
-LAST_INITIALIZE = (0.0, 0.0, 0.0, 0.0, 0.0, 0.0)
-USER_NOTIFY_INTERVAL_S = 10.0
-FREQUENCIES = [1, 2, 4, 10, 20, 50, 100]
-SAMPLING_FREQUENCIES = [2000000, 1000000, 500000, 200000, 100000, 50000,
-                        20000, 10000, 5000, 2000, 1000, 500, 200,
-                        100, 50, 20, 10]
-FLOAT_MAX = np.finfo(float).max
-
-
-def now_str():
-    d = datetime.datetime.utcnow()
-    s = d.strftime('%Y%m%d_%H%M%S')
-    return s
 
 
 def downsample_type_check(string):
@@ -160,21 +145,16 @@ def get_parser():
                         'per minute.')
     p.add_argument('--frequency', '-f',
                    default=2,
-                   choices=FREQUENCIES,
                    type=int,
-                   help='The base collection frequency in Hz.  Defaults to 2.')
-    p.add_argument('--jls', '-j',
-                   type=int,
-                   choices=SAMPLING_FREQUENCIES,
-                   help='Store JLS file with downsampling frequency.')
-    p.add_argument('--source', '-s',
-                   default='stream_buffer',
-                   choices=['stream_buffer', 'sensor'],
-                   help='Select the source for the accumulator data.')
-    p.add_argument('--time',
-                   type=int,
-                   default=None,
-                   help='Enter logging time in seconds.')
+                   help='The base collection frequency in Hz.  Fixed to 2 for JS110.')
+    p.add_argument('--duration',
+                   type=time64.duration_to_seconds,
+                   help='The capture duration in float seconds. '
+                   + 'Add a suffix for other units: s=seconds, m=minutes, h=hours, d=days')
+    p.add_argument('--time-format', '--time_format',
+                   default='utc',
+                   choices=['utc', 'relative'],
+                   help='The time column format.')
     return p
 
 
@@ -189,65 +169,40 @@ def _find_files():
 class Logger:
     """The downsampling logger instance.
 
-    :param header: The CSV header format which is one of ['none', 'simple', 'comment']
-    :param downsample: The downsampling factor in samples.
-        1 performs no downsampling.
-        120 records one sample per minute.
-        None (default) is equivalent to 1.
-    :param frequency: The base frequency, which defaults to 2 Hz.
-    :param resume: Use False or None to start a new logging session.
-        Provide True to resume the most recent logging session.
-    :param jls_sampling_frequency: The sampling frequency for storing JLS files.
-        None (default) does not store a JLS file.
-    :param source: The statistic data source.
+    :param args: The parsed arguments.
     """
-    def __init__(self, header=None, downsample=None, frequency=None, resume=None,
-                 jls_sampling_frequency=None, source=None, test_time=None):
+    def __init__(self, args):
+        self.header = args.header
+        self.resume = bool(args.resume)
+        self.downsample = args.downsample
+        self.frequency = args.frequency
+        self._duration = args.duration
+        self.time_format = args.time_format
         self._start_time_s = None
         self._f_event = None
         self._time_str = None
         self._quit = None
-        self._downsample = 1 if downsample is None else int(downsample)
-        self._frequency = 2 if frequency is None else int(frequency)
+        self.driver = None
         self.log = logging.getLogger(__name__)
-        self._devices = []
+        self._devices = {}  # device_topic -> LoggerDevice
         self._user_notify_time_last = 0.0
-        self._faults = []
-        self._resume = bool(resume)
-        self._jls_sampling_frequency = jls_sampling_frequency
-        self._source = source
-        self._header = header
-        self._base_filename = None
-        self._test_time = test_time
-        if self._frequency not in FREQUENCIES:
-            raise ValueError(f'Unsupported frequency {self._frequency}')
+        self._resync = queue.Queue()
+        self.base_filename = None
+        self._thread_id = threading.current_thread()
 
     def __str__(self):
         return f'Logger("{self._time_str}")'
 
-    def __enter__(self):
-        self.open()
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        try:
-            self.close()
-        except Exception:
-            self.log.exception('While closing during __exit__')
-
-    def _devices_create(self, device_strs):
-        for device_str in device_strs:
-            self._devices.append(LoggerDevice(self, device_str))
-
-    def _on_resume(self):
+    def _on_resume(self, devices_expected):
+        devices = []
         flist = _find_files()
         if not len(flist):
             print('resume specified, but no existing logs found')
-            return
+            return False
         fname = flist[-1]
         base_filename, _ = os.path.splitext(fname)
-        self._base_filename = os.path.join(BASE_PATH, base_filename)
-        event_filename = self._base_filename + '.txt'
+        self.base_filename = os.path.join(BASE_PATH, base_filename)
+        event_filename = self.base_filename + '.txt'
         print('Resuming ' + event_filename)
         with open(event_filename, 'rt') as f:
             for line in f:
@@ -259,324 +214,295 @@ class Logger:
                     elif name == 'start_str':
                         self._time_str = value
                     elif name == 'downsample':
-                        self._downsample = int(value)
+                        self.downsample = int(value)
                     elif name == 'frequency':
-                        self._frequency = int(value)
-                    elif name == 'jls_sampling_frequency':
-                        if value in [None, 'None']:
-                            value = None
-                        else:
-                            value = int(value)
-                        self._jls_sampling_frequency = value
-                    elif name == 'source':
-                        self._source = value
+                        self.frequency = int(value)
+                    elif name == 'duration':
+                        self._duration = None if value == str(None) else float(value)
+                    elif name == 'time_format':
+                        self.time_format = value
+                    else:
+                        print(f'PARAM skip {name}')
                 if 'DEVICES ' in line:
-                    device_strs = line.split(' DEVICES : ')[-1].split(',')
-                    self._devices_create(device_strs)
+                    devices = sorted(line.split(' DEVICES : ')[-1].split(','))
                 if 'LOGGER : RUN' in line:
                     break
         self._f_event = open(event_filename, 'at')
         self.on_event('LOGGER', 'RESUME')
+        if devices_expected is not None and devices != devices_expected:
+            self.on_event('LOGGER', f'RESUME_MISMATCH: expected {devices_expected}, found {devices}')
+        return True
+
+    def _on_new(self, devices):
+        self._start_time_s = time.time()
+        self._time_str = datetime.datetime.now(datetime.UTC).strftime('%Y%m%d_%H%M%S')
+        base_filename = 'jslog_%s_%s' % (self._time_str.split('.')[0], os.getpid())
+        self.base_filename = os.path.join(BASE_PATH, base_filename)
+        event_filename = self.base_filename + '.txt'
+        self._f_event = open(event_filename, 'at')
+        self.on_event('LOGGER', 'OPEN')
+        self.on_event('PARAM', f'start_time={self._start_time_s}')
+        self.on_event('PARAM', f'start_str={self._time_str}')
+        self.on_event('PARAM', f'downsample={self.downsample}')
+        self.on_event('PARAM', f'frequency={self.frequency}')
+        self.on_event('PARAM', f'duration={self._duration}')
+        self.on_event('PARAM', f'time_format={self.time_format}')
+        self.on_event('DEVICES', ','.join(devices))
 
     def open(self):
         self._quit = None
         os.makedirs(BASE_PATH, exist_ok=True)
-
-        if self._resume:
-            self._on_resume()
+        self.driver = Driver()
+        devices = sorted(self.driver.device_paths())
+        self.log.info('Found %d Joulescopes', len(devices))
+        print('Found ' + joulescope_count_to_str(len(devices)))
+        if self.resume and self._on_resume(devices):
+            pass
         else:
-            self._start_time_s = time.time()
-            self._time_str = now_str()
-            base_filename = 'jslog_%s_%s' % (self._time_str, os.getpid())
-            self._base_filename = os.path.join(BASE_PATH, base_filename)
-            event_filename = self._base_filename + '.txt'
-            self._f_event = open(event_filename, 'at')
-            self.on_event('LOGGER', 'OPEN')
-            self.on_event('PARAM', f'start_time={self._start_time_s}')
-            self.on_event('PARAM', f'start_str={self._time_str}')
-            self.on_event('PARAM', f'downsample={self._downsample}')
-            self.on_event('PARAM', f'frequency={self._frequency}')
-            self.on_event('PARAM', f'jls_sampling_frequency={self._jls_sampling_frequency}')
-            self.on_event('PARAM', f'source={self._source}')
+            self._on_new(devices)
 
-            devices = joulescope.scan()
-            device_strs = [str(device) for device in devices]
-            if not len(device_strs):
-                raise RuntimeError('No Joulescopes found')
-            self.log.info('Found %d Joulescopes', len(device_strs))
-            self.on_event('DEVICES', ','.join(device_strs))
-            print('Found ' + joulescope_count_to_str(len(device_strs)))
-            self._devices_create(device_strs)
+        for device in devices:
+            self._devices[device] = LoggerDevice(self, device, self._start_time_s)
+        self.driver.subscribe('@/!add', 'pub', self._on_device_added)
+        self.driver.subscribe('@/!remove', 'pub', self._on_device_removed)
+
+    def _on_device_added(self, topic, value):
+        self._resync.put(('device_added', value))
+
+    def _on_device_removed(self, topic, value):
+        self._resync.put(('device_removed', value))
+
+    def _on_resync(self, resync):
+        cmd, value = resync
+        if cmd == 'event':
+            self._handle_event(*value)
+        elif cmd == 'device_added':
+            if value in self._devices:
+                return
+            self._devices[value] = LoggerDevice(self, value, self._start_time_s)
+        elif cmd == 'device_removed':
+            if value in self._devices:
+                self._devices.pop(value).close()
+        else:
+            print(f'Unsupported command {cmd}')
 
     def close(self):
         self.on_event('LOGGER', 'CLOSE')
-        while len(self._devices):
-            self._devices.pop().close()
+        print(f'duration={time.time() - self._start_time_s:.1f} seconds')
+        for device_topic in list(self._devices.keys()):
+            msg = self._devices.pop(device_topic).close()
+            print(msg)
+            self.on_event('SUMMARY', msg)
 
-        if self._f_event is not None:
-            self.on_event('LOGGER', 'DONE')
-            self._f_event.close()
-            self._f_event = None
+        driver, self.driver = self.driver, None
+        if driver is not None:
+            driver.finalize()
+
+        self.on_event('LOGGER', 'DONE')
+        f_event, self._f_event = self._f_event, None
+        if f_event is not None:
+            f_event.close()
 
     def on_quit(self, *args, **kwargs):
         self.on_event('SIGNAL', 'SIGINT QUIT')
-        self._quit = 'quit from SIGINT'
 
-    def on_stop(self, device, event=0, message=''):
-        # may be called from the Joulescope device thread, keep short!
-        self.on_event('SIGNAL', f'STOP ' + str(device))
-        if event is not None and event > 0:
-            self._faults.append((event, message))
-            self._faults.append(('DEVICE_CLOSE', device))
+    def _handle_event(self, name, message):
+        t_now = datetime.datetime.now(datetime.UTC).isoformat()
+        s = f'{t_now} {name} : {message}\n'
+        self._f_event.write(s)
+        self._f_event.flush()
+        if self.header in ['full', 'comment']:
+            for device in self._devices.values():
+                device.write(f'#& {s}')
+        if name == 'SIGNAL':
+            self._quit = 'quit from SIGNAL'
 
     def on_event(self, name, message):
-        if self._f_event is not None:
-            d = datetime.datetime.utcnow()
-            s = d.strftime('%Y%m%d_%H%M%S.%f')
-            s = f'{s} {name} : {message}\n'
-            self._f_event.write(s)
-            self._f_event.flush()
-            if self._header in ['full', 'comment']:
-                for device in self._devices:
-                    device.write(f'#& {s}')
-
-    def on_event_cbk(self, device, event=0, message=''):
-        # called from the Joulescope device thread, keep short!
-        self._faults.append((event, message))
-        self._faults.append(('DEVICE_CLOSE', device))
-
-    def _open_devices(self, do_notify=True):
-        closed_devices = [d for d in self._devices if not d.is_open]
-        closed_count = len(closed_devices)
-        if closed_count:
-            if do_notify:
-                time_now = time.time()
-                if (self._user_notify_time_last + USER_NOTIFY_INTERVAL_S) <= time_now:
-                    self.log.warning('Missing ' + joulescope_count_to_str(closed_count))
-                    self._user_notify_time_last = time_now
-
-            all_devices = joulescope.scan()
-            for closed_device in closed_devices:
-                for all_device in all_devices:
-                    if str(closed_device) == str(all_device):
-                        closed_device.open(all_device)
-                        break
+        if self._f_event is None:
+            return
+        if self._thread_id == threading.current_thread():
+            self._handle_event(name, message)
         else:
-            self._user_notify_time_last = 0.0
-        return closed_count
+            self._resync.put(('event', (name, message)))
 
     def run(self):
-        self.on_event('LOGGER', 'RUN')
+        self.on_event('LOGGER', 'RUN_START')
         signal.signal(signal.SIGINT, self.on_quit)
+        self.open()
         try:
-            self._open_devices(do_notify=False)
             while not self._quit:
-                closed_count = self._open_devices()
-                if self._test_time is not None:
-                    if time.time() - self._start_time_s > self._test_time:
-                        self._quit = True
-                if closed_count:
-                    time.sleep(0.25)
-                time.sleep(0.1)  # data is received on device's thread
-                if self._source == 'sensor':
-                    for device in self._devices:
-                        device.status()
-                while len(self._faults):  # handle faults on our thread
-                    event, message = self._faults.pop(0)
-                    self.on_event('EVENT', f'{event} {message}')
-                    if event == 'DEVICE_CLOSE':
-                        message.close()
-            for device in self._devices:
-                msg = device.stop()
-                self.on_event('SUMMARY', msg)
+                duration = time.time() - self._start_time_s
+                if self._duration is not None and duration > self._duration:
+                    self._quit = True
+                    continue
+                try:
+                    resync = self._resync.get(timeout=0.05)
+                    self._on_resync(resync)
+                except queue.Empty:
+                    pass
         except Exception as ex:
             self.log.exception('while capturing data')
             self.on_event('FAIL', str(ex))
             return 1
+        finally:
+            self.on_event('LOGGER', 'RUN_DONE')
+            self.close()
 
 
 class LoggerDevice:
 
-    def __init__(self, parent, device_str):
-        self.is_open = False
-        self._parent = weakref.ref(parent)
-        self._device_str = device_str
-        self._device = None
+    def __init__(self, parent: Logger, device: str, start_time_s: float):
+        self._subscribes = []
+        self._parent: Logger = parent
+        self.device = device
+        self.name = device
+        self._start_t64 = time64.as_time64(start_time_s)
         self._f_csv = None
-        self._jls_recorder = None
+        self._t_first = None
 
+        self._stats_count = 0
         self._last = None  # (all values in csv)
-        self._offset = [0.0, 0.0, 0.0]  # [time, charge, energy]
+        self._offset = [0.0, 0.0]  # [charge, energy]
         self._downsample_counter = 0
         self._downsample_state = {
-            'µ': np.zeros(3, dtype=float),
+            'avg': np.zeros(3, dtype=float),
             'min': np.zeros(1, dtype=float),
             'max': np.zeros(1, dtype=float),
         }
         self._downsample_state_reset()
 
+        driver_str, model, sn = self.device.split('/')
+        self.name = f'{model}_{sn}'
+        fname = self._parent.base_filename + f'_{self.name}.csv'
+        parent.on_event('DEVICE', 'OPEN ' + self.device)
+        if os.path.isfile(fname):
+            sz = os.path.getsize(fname)
+            self._last = LAST_INITIALIZE
+            with open(fname, 'rt') as f:
+                f.seek(max(0, sz - CSV_SEEK))
+                for line in f.readlines()[-1::-1]:
+                    if line.startswith('#'):
+                        continue
+                    last_line = line.strip().split(',')
+                    self._last = tuple([last_line[0]] + [float(x) for x in last_line[1:]])
+                    self._offset = [self._last[-2], self._last[-1]]
+                    break
+        self._f_csv = open(fname, 'at+')
+        d = parent.driver
+        d.open(device, mode='restore')
+        if model == 'js110':
+            if parent.frequency != 2:
+                print(f'JS110 only supports frequency 2.  Ignoring frequency {parent.frequency}')
+            d.publish(f'{device}/s/i/range/select', 'auto')
+            self.subscribe(f'{device}/s/sstats/value', 'pub', self._on_statistics)
+        elif model == 'js220':
+            d.publish(f'{device}/s/i/range/mode', 'auto')
+            d.publish(f'{device}/s/v/range/mode', 'auto')
+            scnt = int(round(1_000_000 / parent.frequency))
+            d.publish(f'{device}/s/stats/scnt', scnt)
+            d.publish(f'{device}/s/stats/ctrl', 1)
+            self.subscribe(f'{device}/s/stats/value', 'pub', self._on_statistics)
+        else:
+            print(f'Skip unsupported device {device}')
+
+    def __str__(self):
+        return self.device
+
+    def subscribe(self, topic, flags, fn):
+        self._parent.driver.subscribe(topic, flags, fn)
+        self._subscribes.append((topic, flags, fn))
+
     def _downsample_state_reset(self):
-        self._downsample_state['µ'][:] = 0.0
+        self._downsample_state['avg'][:] = 0.0
         self._downsample_state['min'][:] = FLOAT_MAX
         self._downsample_state['max'][:] = -FLOAT_MAX
 
-    def __str__(self):
-        return self._device_str
-
-    @property
-    def csv_filename(self):
-        sn = self._device_str.split(':')[-1]
-        return self._parent()._base_filename + '_' + sn + '.csv'
-
-    def _resume(self):
-        fname = self.csv_filename
-        if not os.path.isfile(fname):
-            return
-        sz = os.path.getsize(fname)
-        self._last = LAST_INITIALIZE
-        with open(fname, 'rt') as f:
-            f.seek(max(0, sz - CSV_SEEK))
-            for line in f.readlines()[-1::-1]:
-                if line.startswith('#'):
-                    continue
-                self._last = tuple([float(x) for x in line.strip().split(',')])
-                self._offset = [0.0, self._last[-2], self._last[-1]]
-                return
-
-    def open(self, device):
-        if self.is_open:
-            return
-        if str(device) != str(self):
-            raise ValueError('Mismatch device')
-        parent = self._parent()
-        parent.on_event('DEVICE', 'OPEN ' + self._device_str)
-        self._resume()
-        self._f_csv = open(self.csv_filename, 'at+')
-        f = parent._frequency
-        source = parent._source
-        device.parameter_set('reduction_frequency', f'{f} Hz')
-        jls_sampling_frequency = parent._jls_sampling_frequency
-        if jls_sampling_frequency is not None:
-            device.parameter_set('sampling_frequency', jls_sampling_frequency)
-        device.parameter_set('buffer_duration', 2)
-        device.open(event_callback_fn=self.on_event_cbk)
-        if jls_sampling_frequency is not None:
-            time_str = now_str()
-            base_filename = 'jslog_%s_%s.jls' % (time_str, device.device_serial_number)
-            filename = os.path.join(BASE_PATH, base_filename)
-            self._jls_recorder = DataRecorder(filename, calibration=device.calibration)
-            device.stream_process_register(self._jls_recorder)
-        info = device.info()
-        self._parent().on_event('DEVICE_INFO', json.dumps(info))
-        device.statistics_callback_register(self.on_statistics, source)
-        device.parameter_set('i_range', 'auto')
-        device.parameter_set('v_range', '15V')
-        if source == 'stream_buffer' or jls_sampling_frequency is not None:
-            device.parameter_set('source', 'raw')
-            device.start(stop_fn=self.on_stop)
-        self._device = device
-        self.is_open = True
-        return self
-
-    def stop(self):
+    def close(self):
+        for topic, flags, fn in self._subscribes:
+            self._parent.driver.unsubscribe(topic, fn)
+        self._subscribes.clear()
+        self._parent.on_event('DEVICE', 'CLOSE ' + self.device)
+        if self._last is None:
+            msg = f'{self.name}: no data'
+        else:
+            msg = f'{self.name}: charge={self._last[4]:g}, energy={self._last[5]:g}'
+        self._last = None
         try:
-            if self._device is not None:
-                self._device.stop()
-            if self._jls_recorder is not None:
-                self._device.stream_process_unregister(self._jls_recorder)
-                self._jls_recorder.close()
-                self._jls_recorder = None
+            self._parent.driver.close(self.device)
         except Exception:
             pass
-        if self._last is None:
-            msg = f'{self._device} : duration={self.duration:.0f}'
-        else:
-            charge, energy = self._last[4], self._last[5]
-            msg = f'{self._device} : duration={self.duration:.0f}, charge={charge:g}, energy={energy:g}'
-        print(msg)
+        f_csv, self._f_csv = self._f_csv, None
+        if f_csv is not None:
+            f_csv.close()
         return msg
-
-    def on_event_cbk(self, event=0, message=''):
-        self._parent().on_event_cbk(self, event, message)
-
-    def on_stop(self, event=0, message=''):
-        self._parent().on_stop(self, event, message)
-
-    def close(self):
-        self._last = None
-        self.is_open = False
-        if self._device is not None:
-            self._device.close()
-            self._device = None
-        if self._f_csv is not None:
-            self._f_csv.close()
-            self._f_csv = None
 
     def write(self, text):
         if self._f_csv is not None:
             self._f_csv.write(text)
 
-    def status(self):
-        return self._device.status()
-
-    @property
-    def duration(self):
-        now = time.time()
-        t = now - self._parent()._start_time_s + self._offset[0]
-        return t
-
-    def on_statistics(self, data):
+    def _on_statistics(self, topic, value):
         """Process the next Joulescope downsampled 2 Hz data.
 
-        :param data: The Joulescope statistics data.
+        :param value: The Joulescope statistics data.
             See :meth:`joulescope.View.statistics_get` for details.
         """
         # called from the Joulescope device thread
-        parent = self._parent()
+        parent = self._parent
         if self._last is None:
             self._last = LAST_INITIALIZE
 
             columns = ['time', 'current', 'voltage', 'power', 'charge',
                        'energy', 'current_min', 'current_max']
             units = ['s',
-                     data['signals']['current']['µ']['units'],
-                     data['signals']['voltage']['µ']['units'],
-                     data['signals']['power']['µ']['units'],
-                     data['accumulators']['charge']['units'],
-                     data['accumulators']['energy']['units'],
-                     data['signals']['current']['µ']['units'],
-                     data['signals']['current']['µ']['units'],
+                     value['signals']['current']['avg']['units'],
+                     value['signals']['voltage']['avg']['units'],
+                     value['signals']['power']['avg']['units'],
+                     value['accumulators']['charge']['units'],
+                     value['accumulators']['energy']['units'],
+                     value['signals']['current']['avg']['units'],
+                     value['signals']['current']['avg']['units'],
                      ]
             columns_csv = ','.join(columns)
             units_csv = ','.join(units)
             parent.on_event('PARAM', f'columns={columns_csv}')
             parent.on_event('PARAM', f'units={units_csv}')
-            if parent._header in ['simple']:
+            if parent.header in ['simple']:
                 self._f_csv.write(f'{columns_csv}\n')
-            elif parent._header in ['comment', 'full']:
+            elif parent.header in ['comment', 'full']:
                 self._f_csv.write(f'#= header={columns_csv}\n')
                 self._f_csv.write(f'#= units={units_csv}\n')
                 self._f_csv.write(f'#= start_time={parent._start_time_s}\n')
                 self._f_csv.write(f'#= start_str={parent._time_str}\n')
             self._f_csv.flush()
-        t = self.duration
-        i = data['signals']['current']['µ']['value']
-        v = data['signals']['voltage']['µ']['value']
-        p = data['signals']['power']['µ']['value']
-        c = data['accumulators']['charge']['value'] + self._offset[1]
-        e = data['accumulators']['energy']['value'] + self._offset[2]
-        i_min = data['signals']['current']['min']['value']
-        i_max = data['signals']['current']['max']['value']
-        self._downsample_state['µ'] += [i, v, p]
+        t = value['time']['utc']['value'][-1]
+        if self._parent.time_format == 'utc':
+            t = time64.as_datetime(t).strftime('%Y%m%dT%H%M%S.%fZ')
+        else:
+            t = f'{(t - self._start_t64) / time64.SECOND:.3f}'
+        i = value['signals']['current']['avg']['value']
+        v = value['signals']['voltage']['avg']['value']
+        p = value['signals']['power']['avg']['value']
+        c = value['accumulators']['charge']['value']
+        e = value['accumulators']['energy']['value']
+        if self._stats_count == 0:
+            self._offset[0] -= c
+            self._offset[1] -= e
+        c += self._offset[0]
+        e += self._offset[1]
+        i_min = value['signals']['current']['min']['value']
+        i_max = value['signals']['current']['max']['value']
+        self._stats_count += 1
+
+        self._downsample_state['avg'] += [i, v, p]
         self._downsample_state['min'] = np.minimum([i_min], self._downsample_state['min'])
         self._downsample_state['max'] = np.maximum([i_max], self._downsample_state['max'])
         self._downsample_counter += 1
-        if self._downsample_counter >= parent._downsample:
-            s = self._downsample_state['µ'] / self._downsample_counter
+        if self._downsample_counter >= parent.downsample:
+            s = self._downsample_state['avg'] / self._downsample_counter
             self._downsample_counter = 0
             self._last = (t, *s, c, e, *self._downsample_state['min'], *self._downsample_state['max'])
             self._downsample_state_reset()
-            self._f_csv.write('%.7f,%g,%g,%g,%.4f,%g,%g,%g\n' % self._last)
+            self._f_csv.write('%s,%g,%g,%g,%.4f,%g,%g,%g\n' % self._last)
             self._f_csv.flush()
 
 
@@ -587,11 +513,9 @@ def run():
         format='%(asctime)s %(levelname)-8s %(message)s',
         level=logging.WARNING,
         datefmt='%Y-%m-%dT%H:%M:%S')
-    print(f'Starting logging {f"for {args.time} seconds" if args.time is not None else ""} - press CTRL-C to stop')
-    with Logger(header=args.header, downsample=args.downsample,
-                frequency=args.frequency, resume=args.resume,
-                jls_sampling_frequency=args.jls, source=args.source, test_time=args.time) as logger:
-        return logger.run()
+    print(f'Starting logging {f"for {args.time} seconds" if args.duration is not None else ""} - press CTRL-C to stop')
+    logger = Logger(args)
+    return logger.run()
 
 
 if __name__ == '__main__':